feat: add configurable invoice numbering
- Per-seller numbering schemes with template engine ({SEQ}, {MM}, {YYYY}, {DEPT})
- Configurable reset periods: MONTHLY, YEARLY, NEVER
- Atomic UPSERT counters replacing MAX+1 approach (thread-safe)
- Self-healing drift correction: auto-detects and fixes counter/invoice mismatches
- Support for backdated invoices using issue date, not current date
- Flyway migrations V18-V29: schema, backfill, counters, audit, versioning
- Full test coverage: unit, self-healing, concurrency (Testcontainers, disabled by default)

+++ src/main/resources/db/migration/04-invoice-numbering/V19__add_invoice_number_components.sql
@@ -1,11 +1,11 @@
 -- Add Component Columns to Invoices
--- Enables component-based invoice numbering (sortable, resetable)
+-- Enables component-based invoice numbering (sortable, resettable)
 
 ALTER TABLE invoices
 -- Component fields (nullable initially - backfill in V21, NOT NULL in V23)
 ADD COLUMN sequence_number INT COMMENT 'Sequence within period (e.g., 1, 2, 3...)',
-ADD COLUMN period_year SMALLINT COMMENT 'Year component for sorting/reset (e.g., 2026)',
-ADD COLUMN period_month TINYINT DEFAULT 0 COMMENT 'Month: 1-12 for MONTHLY reset, 0 for YEARLY/NEVER',
+ADD COLUMN period_year INT COMMENT 'Year component for sorting/reset (e.g., 2026)',
+ADD COLUMN period_month INT DEFAULT 0 COMMENT 'Month: 1-12 for MONTHLY reset, 0 for YEARLY/NEVER',
 ADD COLUMN department_id BIGINT COMMENT 'Department (v1 feature, FK added in V20)',
 ADD COLUMN invoice_number_display VARCHAR(64) COMMENT 'Human-readable formatted number (e.g., 001-01-2026)';
